Move more helpers to utils.
